# front/openapi.yaml
openapi: 3.0.3
info:
  title: Chat API
  description: AI Chat Application API
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /conversations/{conversationId}/messages:
    get:
      tags:
        - Messages
      summary: 获取对话的所有消息
      operationId: getMessages
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Message'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
    post:
      tags:
        - Messages
      summary: 发送消息
      operationId: sendMessage
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
      responses:
        '201':
          description: 发送成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageResponse'
        '202':
          description: AI 调用失败
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageResponseError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      description: 对话 ID
      schema:
        type: string
...
  schemas:
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        groupConversationId:
          type: string
          format: uuid
        senderId:
          type: string
        senderType:
          type: string
          enum: [user, bot]
        senderName:
          type: string
          description: 发送者名称（群组消息中使用）
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        aiError:
          type: boolean
          default: false
        errorMessage:
          type: string
    MessageResponse:
      type: object
      properties:
        userMessage:
          $ref: '#/components/schemas/Message'
        aiMessage:
          $ref: '#/components/schemas/Message'
    MessageResponseError:
      type: object
      properties:
        userMessage:
          $ref: '#/components/schemas/Message'
        aiMessage:
          type: object
          nullable: true
        error:
          type: object
          properties:
            message:
              type: string
            retryable:
              type: boolean
...